refactor: change solara execution model
We now run an app/code only when the frontend requests it.
This way, we will catch all widgets created, and don't have to ask
for them via the widget protocol (inefficient).
Also, this allows us to run multiple apps, for instance to
make a solara-embed that embed multiple apps into different
parts of the page.