feat: remove <base> tag, use root_path instead.
The base tag causes urls to all be relative to the base tag,
which is not what we want. Instead, we use the root_path to prefix all urls.
Also, the base tag will cause anchor tags to be relative to the base tag,
which will cause the browser to navigate to the wrong url.

Although we can fix this in the javascript side, crawlers will still
navigate to the wrong url.