fix: columns only take up the space they should get
Previously, without overflow: auto, a plotly chart would get too large
and wrap to the next line. Next, it would get really large, and not
shrink anymore. With overflow: auto, it will shrink to the size of the
content, and then scroll if it gets too large.

Content with width=0 will get the minimal amount of size possible.