fix: avoid deadlock on app close in starlette/fastapi
app closing happened in the main thread due to it being an coroutine.
This caused it to execute in the main thread, which tried to get
a lock on a thread in reacton. If reacton is still rendering, it
tried to send a websocket message. Sending this had to go via
the main thread, which was blocked on the lock.