feat: allow reactive fields for lists and dicts to be "invalid"
In practive, a reative field can have a reference to a list or
dict item that can become invalid (e.g. when the list or dict
is changed). We now allow this, and the reactive field
will behave as if it is "disconnected" (e.g. does not receive
change notifications) until the item becomes valid again.