fix: do not call use_reactive on_value when value is updated externally (#76)
It does not make sense to trigger on_value when the value passed
to use_reactive is changed. This has the side effect of the checkbox'
on_value being triggered when the checkbox' value is changed by
the code (it should only be triggered when the user clicks the
checkbox, or the reactive value is changed by some other part of the
code).