fix: add data-base-url in body tag for vegafusion support
Looking at vega-vegafusion, we see it expects the data-base-url tag on
the body tag.

https://github.com/hex-inc/vegafusion/blob/ae2040ebf93b1a69328be9df2ec831bd8fbb6ff4/python/vegafusion-jupyter/src/extension.ts#LL9C17-L9C40

This is present in the widgets-ts-cookiecutter template:
https://github.com/jupyter-widgets/widget-ts-cookiecutter/blob/94f9d384331150655c18a08d94307280c32adedd/%7B%7Bcookiecutter.github_project_name%7D%7D/src/extension.ts#L12

See also:
 * https://github.com/jupyter-widgets/widget-ts-cookiecutter/issues/6

This is required for loading chunked bundles and assets.

We also have this in the voila-vuetify template:

https://github.com/voila-dashboards/voila-vuetify/blame/master/share/jupyter/nbconvert/templates/vuetify-base/index.html.j2#L28

This is the same value we pass ass requirejs' baseUrl:
```
        requirejs.config({
            baseUrl: '{{root_path}}/static/',
```

Fixes https://github.com/widgetti/solara/issues/142