fix: debounce progress bar when kernel busy / computations run
We use the Jupyter kernel busy event and idle to trigger the visualization
of a progress bar to indicate that a computation is happening. However,
it does not make sense to do this for every single computation, as it
would be too much noise. Instead, we debounce the progress bar to only
show it when the kernel is busy for more than 300 msec.

Fixes #143