feat: do not use pydantic
We only use pydantic for the settings, and it is not worth the
dependency. We made the API quite similar, but it is only for internal
use.

We cannot keep using pydantic because v2 does not include pydantic
settings code and needs the pydantic-settings dependency.
However, adding that dependency indirectly requires pydantic v2.