feat: warn user when modules can be loaded twice.
Solara by default assumes the name for `solara run <name>` is a
directory. However, when the name is also a package, which imports
module from the package, the package modules can be loaded twice.
Solara will execute the python files in the directory, while Python
will load the python files as modules from the package.
This can leads to unexpected behavior, such as reactive variables
existing twice.