fix: make vega and altair work in the jupyter notebook
Because the vega url does not end with .js the CdnHandler does not
serve the file with the correct mimetype. We do not store the mime type
on disk, so we do not know the mimetype of the file.
Currently, the mimetype is guessed based on the file extension.
By giving a more explicit path that ends in .js, we can make sure the
mimetype is set correctly.
In the end this caused the Altair component not to work.