test: make test more robust by non using globals and guessing the context
We relied on a global variable to wait for the component to be called,
but this seems to not be very robust. Instead, we pass a unique ID
via a query parameter to not re-use the Event object, and store the
context that we actually use, instead of relying on the last used
context found.