refactor: do not run all runners in ipywidgets_runner fixture (#310)
Each time the ipywidgets_runner fixture was used, all fixtures are used. This means the ipywidgets_runner_solara runner was also used, causing the solara_test fixture to always run and also close the app each time, which close all the contexts.

This is bad for performance of the tests, but it is possibly also the cause for some flakeyness in our tests.