feat: allow reconnecting to existing kernel and display widget by id
This allows an ipypopout or a similar library to open a new browser
window and show a widget that is already running in the main window.

Note that this is limited to the same browser, because the session_id
is required to be the same. This is a security feature.