refactor: replace --dev with --auto-restart and have a --production mode
Before, we had a --dev mode that would auto-restart the server when
files changed. This means that auto restart was connected to dev mode.
Now, by default solara is in development mode, and the --auto-restart
option is separated.
This means that we now have a special --production mode that should
be used in production settings. It will serve users with minimized
JS and CSS and disable project and .vue reloading.

This means --reload (a bad and confusing name) and --dev are now
deprecated, but will behave the same as --auto-reload/-a.