feat: added use_activator_width option for Menu and removed`<div>` around activator (#345)
The extra div can cause layout issues if the activator is a span or has flexbox. Avoiding this extra div causes less flow issues. Also by having use_activator_width configable we can have the menu follow the size of the activator, or not (have the content decide).
In the case of a date picker, we just want it to be the size of what the data picker needs, so we would use `use_activator_width=False`.