fix: avoid re-render when using sidebar or appbar
Before we had a 'fast path' which did not render the nav drawer and
appbar when not used. This caused a different component tree to be
rendered, which caused a re-render of the entire app.

This also causes our parent div to always be the same, which will
make the app behave more consistently (always 100vh parent).

Fixes #331