fix: do not use excessive memory when rendering markdown
Due to legacy reasons we enabled the emoji plugin from pymdown.
Every time the Markdown object gets created, it will load the
emoji extension and will deep copy the emoji database.
This is about 150kb of data:
len(str(pymdownx.emoji1_db.emoji))

This is not a lot, but it is unnecessary and can be avoided.