fix: avoid extra scrollbar when having no appbar or sidebar
This issue was introduced in 3884d6a035ab11098d78f6ff6c76aec060dfd785
and was caused by a full height 100% chain and a single element with
padding, causing the solara-autorouter-content to be pushed down
and creating scrollbars on its parents.
If instead, we add the scrollbar to the solara-autorouter-content
element, we can avoid this issue.