chore: upgrade mypy to 1.6.0
This is required for #461
but upgrading to an even never version will lead to many errors/issues.
This version is the smallest upgrade that fixes the issue for this PR.

Note that was previously done in f7bfb29551b6786a85370a88a3c3b10680f401d1
but due to further changes we need to upgrade again.