feat: improve matplotlib support - mimic Jupyter (#540)
feat: improve matplotlib support - as Jupyter


Several improvements:

## Default backend
We do this by checking if the env var 'MPLBACKEND' is not set, and if so, we set it to
module://matplotlib_inline.backend_inline

## Support for display(figure)

This makes it behave more like the Jupyter notebook environment.

## Support matplotlib pylab interface in solara server

This makes code that 'just works' in the notebook also work in solara
server. We patch global dicts with scoped dicts.

See #539