fix: avoid closing a kernel twice by using a mutex
A kernel could be close multiple times if the websocket was closed at
the same time as the close beacon was send. This caused in CI
KeyError on custom_storage.py when the kernel cleanup
function was called twice.
State managegment for pages is not protected by a mutex, so that
it is thread safe.