fix: catch PermissionError checking if directory exists
jupyter_core.paths.jupyter_path("nbextensions") may return directories
that do not exist. However, the user may not have permission to read
some of these paths, leading to a PermissionError when calling
pathlib.Path.exists(), which crashes the application.

This change catches the PermissionError and logs it as a warning,
preventing the application from crashing. Essentially: if the user does
not have permission to the path, it might as well not exist.

See https://github.com/widgetti/solara/issues/611