fix: support esm define when running under virtual kernel
When we define a esm top level in a module, we don't need to create
the module widget. However, if we do this during the running of the
app under the context of a virtual kernel, we should directly
create the esm widget.

Should fix the failing of https://github.com/widgetti/ipyantd/pull/4