pre-commit: Add auto-update schedule
To make sure the pre-commit configuration stays up to date, add auto-update schedule to the pre-commit configuration file.

pre-commit.ci needs to be enable for this (which I highly recommend). When enabled, it will check once a week if any pre-commit action can be updated, and will open a PR if so.