fix: to redirect back after auth we used the wrong base url
settings.main.base_url should have been used, since that can be
used to override the base url from starlette.

Note that this should always never be needed once a proxy and uvicorn
are configured correctly (see also #745).

But starlette does not respect the x-forwarded-host header, so we
might need to use this in the future in situations of proxies that
need to set the Host header due to https communication.

Fixes #740