feat: allow connecting to a running kernel from a different browser
This is useful for connecting external devices, such as smartphone or
AR devices (Apple Vision pro and Meta Quest) to the running kernel
using ipypopout.

This is normally not possible since we require the connecting client
to have the same session id as the already kernel.

This this setting, we can disable the session check and the above use
case.

e.g.

SOLARA_KERNEL_SESSION_CHECK=False solara run ...

setting to disable session check