docs: fixes for mobile users
Because we previously used a `solara.v.NavigationDrawer` with `v_model=True`, that was only hidden and not destroyed on mobile, Vuetify would think that a floating navigation drawer is open on page load. This would mean that the page could not be used before the user tapped somewhere on the page to "close the drawer". Also adds another maximum width restriction to make sure content doesn't become wider than the viewport on mobile resolutions.