fix: a failing websocket.send would drop the connection silently
Instead of assuming all exceptions are due to a closed connection,
we only ignored the exception when it is due to a closed connection.

This caused issues in #841 where we called from the same thread as the
websocket's portal, which caused the connection to be ignored without
any error message.