fix: do not crash on use close menu
This is more of a workaround than a fix.
But in certain cases (I suspect multiple rerenders) use get_widget
call fails. This workaround did not cause the hook not to work,
which confirms the suspiction.
This is hard to debug, so this is a good enough fix for now.